Unify case statement
The two case statements are meant to allow a single version to be
auto-aliased twice (once with the minimum patch segment removed, and
again with the maximum patch segment removed)

Due to the idempotent behavior of `alias <version> --auto`, running the
same alias --auto twice has no ill effect. Thus, if the shortest-match
and longest-match versions are identical, there's no harm in simply
running alias --auto twice. This reduces a bunch of redundant code that
isn't completely clear in its purpose. (and ran the risk of
inadvertently being merged into a single case with 2 branches, which
would have actually been a regression)