Refactor patch_start_point -> sort_versions
Rather than having patch_start_point function that just returns the
starting character point of the patch version section, what we really
want is a function that simply sorts the versions intelligently.

Create sort_versions function that takes all the versions to sort on
stdin and intelligently sorts using the proper field components of a
version string. In this way, the sort_versions function is really just a
wrapper for `sort` with the added knowledge of which arguments to use
for `sort` and how to sort on the proper fields of a version string. The
old patch_start_point logic is embedded in this function.