Fix autoaliasing semver versions
The point_releases function only matched semver versions of the form
<digit>.<digit>.<digit>. So a version like 2.1.11 wouldn't be matched.

The point_releases function had duplicated the same essential pattern of
echo;sed;sort across the three different styles of version numbers
(X.Y.Z-pNN, jruby-X.Y.Z, X.Y.Z). The breakdown between semver and
jruby-1.2.3 styles was a meaningless distinction because in both cases,
the trailing .* should be trimmed. So instead of artificially separating
these cases, they can be unified to simply strip the final patch segment
regardless of the style of version string. Unifying the sed portion for
each of the version styles means that the glob could be reduced to
*.*.*, which ensures matching against semver versions with >10 segments.